requirements.txt
asgiref==3.6.0
Django==3.2.18
pytz==2023.2
sqlparse==0.4.3
typing_extensions==4.5.0
Pillow==9.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
django-cors-headers==3.14.0
django-filter==23.1
PyJWT==2.6.0

